requirements.txt
python-telegram-bot>=20.0
aiohttp
pip>=24.3.1

